+++ apps/ui/src/widgets/general/index.ts
@@ -1,0 +1,7 @@
+export { isGeneralWidgetType, GENERAL_WIDGET_TYPES, generalDefaultSize, generalDefaultConfig } from "./config";
+export type { GeneralWidgetType } from "./config";
+export { CalendarWidget } from "./CalendarWidget";
+export { WeatherWidget } from "./WeatherWidget";
+export { SystemInfoWidget } from "./SystemInfoWidget";
+export { DeviceStatusWidget } from "./DeviceStatusWidget";
+export { GeneralWidgetBody } from "./GeneralWidgetBody";